since it makes the caclulation much faster, and sinc acoefficient + bcoefficient = 1, it makes sense to have the output be 32 bits and omit the limit check. It makes the whole execution run 20% fater

+++ src/FirstOrderIIRFilter.h
@@ -14,7 +14,7 @@ class FirstOrderIIRFilter{
     void reset();
     
     private:
-    int64_t output;
+    int32_t output;
     int32_t aCoefficient;
     int32_t bCoefficient;
